docker flush warn

+++ project/bash/ubuntu-16.x/docker/flush.sh
@@ -1,6 +1,13 @@
 #!/usr/bin/env bash
 
 dockerFlush() {
+
+  # Warn user.
+  read -p "It will remove all containers / images / networks, are you sure ? Type n to stop : " ANSWER
+  if [ ${ANSWER} == "n" ];then
+    return
+  fi;
+
   wex docker/stopAll
   # Remove all images
   docker rmi $(docker images -qa) -f
